add send_typing_action
